requirements.txt
face-recognition==1.3.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
google-cloud-vision==2.0.0
